fix(figures): stop rewriting a Moonbot object's creation instant
Dragging a Moonbot figure in the terminal made it vanish. Measured, not reasoned
about: the app log shows our upsert at 11:07:50.799, the core echoing it back at
.998, and the object DELETED at 11:07:51.201. Diffing our outgoing blob against
the last one Moonbot sent, every byte matches except `@22` — `dd…` from Moonbot
against `dc…` from us.

That is the creation instant. The wire carries a Delphi `TDateTime`, whose
resolution is finer than a millisecond, and the figure store kept it as whole
milliseconds; a re-upsert therefore wrote back an instant the object never had,
and Moonbot dropped it. Rounding instead of truncating had already narrowed the
error, but the only correct amount to change a field we did not edit is none, so
the store keeps the value to the bit.

Also, the three lines a drag can take hold of — both ends and the free level —
are drawn heavier while the figure is hovered or selected. A full-width line has
no X for a knot to sit at, so weight is the only cue available, and it is the one
Moonbot uses. Weight and not colour, so a level keeps the hue it is recognised by.

+++ crates/moon-core/src/figures/mod.rs
@@ -68,7 +68,13 @@ pub struct Figure {
     #[serde(default)]
     pub line_kind: LineKind,
     /// Creation time in Unix milliseconds, shown in the alert list's Time column.
-    pub created_ms: i64,
+    ///
+    /// Fractional on purpose. The wire carries a Delphi `TDateTime`, whose resolution is finer than
+    /// a millisecond, and this field is what a re-upsert writes back into it. Rounding it to whole
+    /// milliseconds rewrote the creation instant of an object drawn in Moonbot on every drag —
+    /// measured, byte for byte at `@22` — and Moonbot then DELETED the object about 200 ms later.
+    /// A figure drawn here has no fraction to lose, so nothing is paid for keeping it.
+    pub created_ms: f64,
     /// Whether the "Alert" checkbox is enabled and the figure is sent to the core as a chart alert.
     pub alert: bool,
     /// Associated strategy ID of the "Alerts" type; 0 means no strategy. Stored at blob offset 32.
@@ -93,7 +99,7 @@ pub struct Figure {
 
 impl Figure {
     /// Builds a local figure of `kind` in the current drawing style.
-    pub fn new(kind: FigureKind, style: DrawStyle, created_ms: i64) -> Self {
+    pub fn new(kind: FigureKind, style: DrawStyle, created_ms: f64) -> Self {
         Self {
             id: 0,
             kind,
